+++ src/main/java/subway/application/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.application.dto;
 
 import javax.validation.constraints.NotBlank;
