+++ src/main/java/subway/application/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.application.dto;
 
 import java.util.List;
