+++ src/main/java/subway/application/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.application.dto;
 
 import javax.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.Range;
