+++ src/main/java/subway/ui/StationController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.StationService;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.application.dto.StationRequest;
+import subway.application.dto.StationResponse;
 
 @RestController
 @RequestMapping("/stations")
