+++ src/test/java/subway/application/LineServiceTest.java
@@ -33,10 +33,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import subway.domain.line.LineRepository;
 import subway.domain.section.SectionRepository;
 import subway.domain.station.StationRepository;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
-import subway.dto.SectionRequest;
-import subway.dto.StationResponse;
+import subway.application.dto.LineRequest;
+import subway.application.dto.LineResponse;
+import subway.application.dto.SectionRequest;
+import subway.application.dto.StationResponse;
 import subway.exception.BadRequestException;
 import subway.exception.NotFoundException;
