+++ src/test/java/subway/fixture/StationFixture.java
@@ -4,7 +4,7 @@ import java.util.List;
 import subway.dao.entity.StationEntity;
 import subway.domain.station.Station;
 import subway.domain.station.dto.StationRes;
-import subway.dto.StationRequest;
+import subway.application.dto.StationRequest;
 
 public final class StationFixture {
